# public/swagger-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/mail-blueprint/public": {
      "get": {
        "operationId": "MailBlueprintCurdController_findAllPublic",
        "summary": "Get all public Mail Blueprints",
        "parameters": [
          {
            "required": false,
            "name": "query",
            "in": "query",
            "schema": {
              "example": {
                "query": {
                  "select": {
                    "name": true
                  },
                  "where": {
                    "email": "hniqbal01@gmail.com"
                  },
                  "page": 1,
                  "limit": 10
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mail Blueprints retrived successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailBlueprintPaginationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-blueprint"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/mail-blueprint": {
      "get": {
        "operationId": "MailBlueprintCurdController_findAll",
        "summary": "Get all Mail Blueprints",
        "parameters": [
          {
            "required": false,
            "name": "query",
            "in": "query",
            "schema": {
              "example": {
                "query": {
                  "select": {
                    "name": true
                  },
                  "where": {
                    "email": "hniqbal01@gmail.com"
                  },
                  "page": 1,
                  "limit": 10
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mail Blueprints retrived successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailBlueprintPaginationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-blueprint"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "MailBlueprintCurdController_create",
        "summary": "Create a new mailBlueprint",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMailBlueprintDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mail Blueprint created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailBlueprint"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-blueprint"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/mail-blueprint/public/{id}": {
      "get": {
        "operationId": "MailBlueprintCurdController_findByIdPublic",
        "summary": "Get public mailBlueprint by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MailBlueprint ID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Mail Blueprint retrived successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailBlueprint"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-blueprint"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/mail-blueprint/{id}": {
      "get": {
        "operationId": "MailBlueprintCurdController_findById",
        "summary": "Get mailBlueprint by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MailBlueprint ID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Mail Blueprint retrived successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailBlueprint"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-blueprint"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "MailBlueprintCurdController_update",
        "summary": "Update mailBlueprint by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MailBlueprint ID",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMailBlueprintDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mail Blueprint updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailBlueprint"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-blueprint"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "MailBlueprintCurdController_delete",
        "summary": "Delete mailBlueprint by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MailBlueprint ID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Mail Blueprint deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailBlueprint"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-blueprint"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/mail-automation/public": {
      "get": {
        "operationId": "MailAutomationCurdController_findAllPublic",
        "summary": "Get all public mail automations",
        "parameters": [
          {
            "required": false,
            "name": "query",
            "in": "query",
            "schema": {
              "example": {
                "query": {
                  "select": {
                    "name": true
                  },
                  "where": {
                    "email": "hniqbal01@gmail.com"
                  },
                  "page": 1,
                  "limit": 10
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mail Automation retrived successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailAutomationPaginationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-automation",
          "v1/mail-blueprint"
        ]
      }
    },
    "/api/mail-automation": {
      "get": {
        "operationId": "MailAutomationCurdController_findAll",
        "summary": "Get all mail automations",
        "parameters": [
          {
            "required": false,
            "name": "query",
            "in": "query",
            "schema": {
              "example": {
                "query": {
                  "select": {
                    "name": true
                  },
                  "where": {
                    "email": "hniqbal01@gmail.com"
                  },
                  "page": 1,
                  "limit": 10
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mail Automation retrived successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailAutomationPaginationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-automation",
          "v1/mail-blueprint"
        ]
      },
      "post": {
        "operationId": "MailAutomationCurdController_create",
        "summary": "Create a new mailAutomation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMailAutomationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mail Automation created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailAutomation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-automation",
          "v1/mail-blueprint"
        ]
      }
    },
    "/api/mail-automation/public/{id}": {
      "get": {
        "operationId": "MailAutomationCurdController_findByIdPublic",
        "summary": "Get public mailAutomation by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MailAutomation ID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Mail Automation retrived successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailAutomation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-automation",
          "v1/mail-blueprint"
        ]
      }
    },
    "/api/mail-automation/{id}": {
      "get": {
        "operationId": "MailAutomationCurdController_findById",
        "summary": "Get mailAutomation by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MailAutomation ID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Mail Automation retrived successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailAutomation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-automation",
          "v1/mail-blueprint"
        ]
      },
      "put": {
        "operationId": "MailAutomationCurdController_update",
        "summary": "Update mailAutomation by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MailAutomation ID",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMailAutomationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mail Automation updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailAutomation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-automation",
          "v1/mail-blueprint"
        ]
      },
      "delete": {
        "operationId": "MailAutomationCurdController_delete",
        "summary": "Delete mailAutomation by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "MailAutomation ID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Mail Automation deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MailAutomation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExceptionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/mail-automation",
          "v1/mail-blueprint"
        ]
      }
    },
...
  },
  "info": {
    "title": "Boilerplate v1 API",
    "description": "Basic docs for API in version 1",
    "version": "1.0",
    "contact": {
      "name": "Shirsho ",
      "url": "https://marketron.com",
      "email": "hniqbal01@gmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ExceptionResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "Invalid request parameters"
          },
          "timestamp": {
            "type": "string",
            "example": "2024-11-22T04:10:35.729Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/resource"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "timestamp",
          "path"
        ]
      },
...
      "MailBlueprint": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The unique name of the mail blueprint",
            "example": "WelcomeEmail"
          },
          "bodyContent": {
            "type": "string",
            "description": "The content/body of the email",
            "example": "Hello, welcome to our platform!"
          },
          "subjectContent": {
            "type": "string",
            "description": "The subject line of the email",
            "example": "Welcome to Our Platform"
          },
          "placeholders": {
            "description": "List of placeholder keys that can be replaced in the content",
            "example": [
              "username",
              "date"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "bodyContent",
          "subjectContent",
          "placeholders"
        ]
      },
      "MailBlueprintPaginationResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MailBlueprint"
            }
          },
          "count": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "data",
          "count"
        ]
      },
      "CreateMailBlueprintDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Blueprint name"
          },
          "bodyContent": {
            "type": "string",
            "example": "Shirsho {{placeholder}}"
          },
          "subjectContent": {
            "type": "string",
            "example": "Shirsho {{placeholder}}"
          }
        },
        "required": [
          "name",
          "bodyContent",
          "subjectContent"
        ]
      },
      "UpdateMailBlueprintDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Blueprint name"
          },
          "bodyContent": {
            "type": "string",
            "example": "Shirsho {{placeholder}}"
          },
          "subjectContent": {
            "type": "string",
            "example": "Shirsho {{placeholder}}"
          }
        }
      },
      "MailAutomation": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the mail automation",
            "example": "WelcomeEmail"
          },
          "resourceName": {
            "type": "string",
            "description": "The resource name of the mail automation",
            "example": "exam"
          },
          "bluePrintId": {
            "type": "string",
            "description": "The blueprint id of the mail automation",
            "example": "mongo id"
          }
        },
        "required": [
          "resourceName",
          "bluePrintId"
        ]
      },
      "MailAutomationPaginationResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MailAutomation"
            }
          },
          "count": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "data",
          "count"
        ]
      },
      "CreateMailAutomationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Shirsho",
            "description": "Name of the mail automation optional"
          },
          "resourceName": {
            "type": "string",
            "example": "example"
          },
          "bluePrintId": {
            "type": "string",
            "example": "mongo id"
          }
        },
        "required": [
          "resourceName",
          "bluePrintId"
        ]
      },
      "UpdateMailAutomationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Shirsho",
            "description": "Name of the mail automation optional"
          },
          "resourceName": {
            "type": "string",
            "example": "example"
          },
          "bluePrintId": {
            "type": "string",
            "example": "mongo id"
          }
        }
      },
...
    }
  }
}